+++ TalkWithYourFiles/flow_coordinator.py
@@ -7,6 +7,7 @@ from parameter_controller import ParameterController
 
 from typing import List, Optional, Tuple, Dict, IO
 import logging
+from pydantic.error_wrappers import ValidationError
 
 """
 This file, flow_coordinator.py, serves as a central coordination module within the application. It acts as a bridge between user interface & underlying functionalities of other modules.
@@ -61,8 +62,11 @@ class FlowCoordinator:
         """
 
         # # Set up with the configurations.
-        self.runner.setup()
-
+        try:
+            self.runner.setup()
+        except ValidationError as ve:
+            return "Invalid or missing API key. Please ensure you have entered a valid OpenAI API key."
+      
         ## VERIFY THE INPUT BEFORE STARTING WITH THE REST
         is_valid, error_message = self.validate_input(files, user_question)
         if not is_valid:
