+++ TalkWithYourFiles/streamlit_modules/streamlit_helper_functions.py
@@ -96,15 +96,12 @@ def advanced_parameters_section(param_controller):
 def create_authorization_box(flow_coordinator): 
     ##### Authorization & Setting up the environment variable
 
-    ## causes key leak without modifying streamlit caching options.
-    ## risky gate. lets close using dotenv file?
-    # default_key = flow_coordinator.authorizer.get_api_key()
-    
-    default_key = None
-    input_api_key = st.sidebar.text_input("Enter your OpenAI API key", 
-                                value=default_key if default_key else "", 
-                                type="password"
-                                )
+
+    input_api_key = st.sidebar.text_input(
+        "Enter your OpenAI API key", 
+        value="", 
+        type="password"
+    )
     flow_coordinator.authorizer.set_api_key_environment_variable(input_api_key)
